Refactor uploadFiles in a new function
